Moved JS click methods to DesktopWebDriverUtils
