+++ driver-ui/src/main/java/eu/tsystems/mms/tic/testframework/utils/JSUtils.java
@@ -230,10 +230,6 @@ public final class JSUtils {
         return sb.toString();
     }
 
-    private static String toHex(Color color) {
-        return String.format("#%02x%02x%02x%02x", color.getRed(), color.getGreen(), color.getBlue(), color.getAlpha());
-    }
-
     private static boolean isJavascriptImplementedOnPage(final WebDriver driver, final String id) {
         List<WebElement> script = driver.findElements(By.xpath(".//script[@id='" + id + "']"));
         return script.size() > 0;
@@ -546,86 +542,86 @@ public final class JSUtils {
         );
     }
 
-    public void mouseOver(final WebDriver webDriver, final WebElement webElement) {
-        demoMouseOver(webDriver, webElement);
-        pMouseOverJS(webDriver, webElement);
-    }
+//    public void mouseOver(final WebDriver webDriver, final WebElement webElement) {
+//        demoMouseOver(webDriver, webElement);
+//        pMouseOverJS(webDriver, webElement);
+//    }
 
-    public void doubleClick(final WebDriver webDriver, final WebElement webElement) {
-        Point location = webElement.getLocation();
-        JSUtils.executeJavaScriptMouseAction(webDriver, webElement, JSMouseAction.DOUBLE_CLICK, location.getX(), location.getY());
-    }
+//    public void doubleClick(final WebDriver webDriver, final WebElement webElement) {
+//        Point location = webElement.getLocation();
+//        JSUtils.executeJavaScriptMouseAction(webDriver, webElement, JSMouseAction.DOUBLE_CLICK, location.getX(), location.getY());
+//    }
 
-    public void rightClick(final WebDriver webDriver, final WebElement webElement) {
-        String script = "var element = arguments[0];" +
-                "var e = element.ownerDocument.createEvent('MouseEvents');" +
-                "e.initMouseEvent('contextmenu', true, true,element.ownerDocument.defaultView, 1, 0, 0, 0, 0, false,false, false, false,2, null);" +
-                "return !element.dispatchEvent(e);";
+//    public void rightClick(final WebDriver webDriver, final WebElement webElement) {
+//        String script = "var element = arguments[0];" +
+//                "var e = element.ownerDocument.createEvent('MouseEvents');" +
+//                "e.initMouseEvent('contextmenu', true, true,element.ownerDocument.defaultView, 1, 0, 0, 0, 0, false,false, false, false,2, null);" +
+//                "return !element.dispatchEvent(e);";
+//
+//        JSUtils.executeScript(webDriver, script, webElement);
+//    }
 
-        JSUtils.executeScript(webDriver, script, webElement);
-    }
+//    public void click(final WebDriver webDriver, final WebElement webElement) {
+//        executeScript(webDriver, "arguments[0].click();", webElement);
+//    }
 
-    public void click(final WebDriver webDriver, final WebElement webElement) {
-        executeScript(webDriver, "arguments[0].click();", webElement);
-    }
+//    private void demoMouseOver(final WebDriver webDriver, final WebElement webElement) {
+//        if (Testerra.Properties.DEMO_MODE.asBool()) {
+//            UiElementHighlighter elementHighlighter = Testerra.getInjector().getInstance(UiElementHighlighter.class);
+//            elementHighlighter.highlight(webDriver, webElement, new Color(255, 255, 0));
+//        }
+//    }
 
-    private void demoMouseOver(final WebDriver webDriver, final WebElement webElement) {
-        if (Testerra.Properties.DEMO_MODE.asBool()) {
-            UiElementHighlighter elementHighlighter = Testerra.getInjector().getInstance(UiElementHighlighter.class);
-            elementHighlighter.highlight(webDriver, webElement, new Color(255, 255, 0));
-        }
-    }
+//    public void clickAbsolute(final WebDriver webDriver, final WebElement webElement) {
+//        pClickAbsolute(webDriver, webElement);
+//    }
 
-    public void clickAbsolute(final WebDriver webDriver, final WebElement webElement) {
-        pClickAbsolute(webDriver, webElement);
-    }
+//    public void mouseOverAbsolute2Axis(final WebDriver webDriver, final WebElement webElement) {
+//        demoMouseOver(webDriver, webElement);
+//        pMouseOverAbsolute2Axis(webDriver, webElement);
+//    }
 
-    public void mouseOverAbsolute2Axis(final WebDriver webDriver, final WebElement webElement) {
-        demoMouseOver(webDriver, webElement);
-        pMouseOverAbsolute2Axis(webDriver, webElement);
-    }
+//    private void pMouseOverJS(final WebDriver webDriver, final WebElement webElement) {
+//        final String code = "var fireOnThis = arguments[0];"
+//                + "var evObj = document.createEvent('MouseEvents');"
+//                + "evObj.initEvent( 'mouseover', true, true );"
+//                + "fireOnThis.dispatchEvent(evObj);";
+//
+//        ((JavascriptExecutor) webDriver).executeScript(code, webElement);
+//    }
 
-    private void pMouseOverJS(final WebDriver webDriver, final WebElement webElement) {
-        final String code = "var fireOnThis = arguments[0];"
-                + "var evObj = document.createEvent('MouseEvents');"
-                + "evObj.initEvent( 'mouseover', true, true );"
-                + "fireOnThis.dispatchEvent(evObj);";
+//    private void pClickAbsolute(WebDriver driver, WebElement webElement) {
+//        // Start the StopWatch for measuring the loading time of a Page
+//        StopWatch.startPageLoad(driver);
+//
+//        Point point = webElement.getLocation();
+//
+//        Actions action = new Actions(driver);
+//
+//        // goto 0,0
+//        action.moveToElement(webElement, 1 + -point.getX(), 1 + -point.getY());
+//
+//        // move y, then x
+//        action.moveByOffset(0, point.getY()).moveByOffset(point.getX(), 0);
+//
+//        // move to webElement
+//        action.moveToElement(webElement);
+//        action.moveByOffset(1, 1);
+//        action.click().perform();
+//    }
 
-        ((JavascriptExecutor) webDriver).executeScript(code, webElement);
-    }
-
-    private void pClickAbsolute(WebDriver driver, WebElement webElement) {
-        // Start the StopWatch for measuring the loading time of a Page
-        StopWatch.startPageLoad(driver);
-
-        Point point = webElement.getLocation();
-
-        Actions action = new Actions(driver);
-
-        // goto 0,0
-        action.moveToElement(webElement, 1 + -point.getX(), 1 + -point.getY());
-
-        // move y, then x
-        action.moveByOffset(0, point.getY()).moveByOffset(point.getX(), 0);
-
-        // move to webElement
-        action.moveToElement(webElement);
-        action.moveByOffset(1, 1);
-        action.click().perform();
-    }
-
-    private void pMouseOverAbsolute2Axis(WebDriver driver, WebElement webElement) {
-        Actions action = new Actions(driver);
-
-        Point point = webElement.getLocation();
-
-        // goto 0,0
-        action.moveToElement(webElement, 1 + -point.getX(), 1 + -point.getY()).perform();
-
-        // move y, then x
-        action.moveByOffset(0, point.getY()).moveByOffset(point.getX(), 0).perform();
-
-        // move to webElement
-        action.moveToElement(webElement).perform();
-    }
+//    private void pMouseOverAbsolute2Axis(WebDriver driver, WebElement webElement) {
+//        Actions action = new Actions(driver);
+//
+//        Point point = webElement.getLocation();
+//
+//        // goto 0,0
+//        action.moveToElement(webElement, 1 + -point.getX(), 1 + -point.getY()).perform();
+//
+//        // move y, then x
+//        action.moveByOffset(0, point.getY()).moveByOffset(point.getX(), 0).perform();
+//
+//        // move to webElement
+//        action.moveToElement(webElement).perform();
+//    }
 }
